requirements.txt
# Slack SDK
slack-bolt>=1.18.0
slack-sdk>=3.27.0
aiohttp>=3.9.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Strands Agents
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AgentCore Runtime
bedrock-agentcore>=0.1.0

# Database
pymysql>=1.1.0
pymysqlpool>=0.3.7

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0

# Date/Time parsing
python-dateutil>=2.8.0

